third_party/blink/web_tests/compositing/select-element-expected.txt

This is a testharness.js-based test.
[FAIL] scrollable
  assert_equals: [object Object] expected 4 but got 1
Harness: the test ran to completion.
